backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
firebase-admin
python-dotenv
pydantic
pydantic-settings
pyjwt
passlib
python-decouple
aiomqtt
loguru
